BUGFIX & improvement : new biblioitems fields saving and GetMarcAuthors improvement
hdl defined 5 new fields in biblioitems, to handle collections.
The _koha_modify_biblioitem didn't deal with those fields. it's fixed now.

GetMarcAuthor in UNIMARC have a $4 field that is related to the responsability the author has
it is connected to an authorised value.
The improvement expand the numeric value to the expanded one.

Signed-off-by: Chris Cormack <crc@liblime.com>