bump to 3.00.00.006 : issues.issuedate auto fill added on update
In Koha 2.2 there was no field to store the issue date. There was only a timestamp
which is modified every time you do something on the line.
it's fixed in 3.0, but the new issuedate field is empty for 2.2 catalogue migrated to 3.0
This fills the issue date with a value that is probably wrong, but we can't know the correct one.
at least we have something and the stats works (Date::Calc is very unhappy if the field is empty)

Signed-off-by: Chris Cormack <crc@liblime.com>
Signed-off-by: Joshua Ferraro <jmf@liblime.com>