temp bugfix -- restore file upload functionality
CGI->new() should not be used in any C4 module -
it has a side-effect that breaks the AJAX
file upload script.  Specifically, any file upload
field is slurped when a new CGI object is created, and
is not available to a second object created by CGI->new().

Consequently, C4::Debug should not be instantiating
a CGI object in its BEGIN block.  It can use other CGI::* modules
to parse QUERY_STRING to get the value of the debug
parameter (note that this means that the debug param
can be reliably set only in a GET, not as part of a
form submission).

Signed-off-by: Chris Cormack <crc@liblime.com>
Signed-off-by: Joshua Ferraro <jmf@liblime.com>