bug 2083: requiring that the test database have a different name from the production database
This patch requires that the test database name be different from the production database name.
The object here is to prevent people from specifying the same database for both, and then
losing all of their data in the production database.

This may cause problems for people who want to do that, though I consider that bad practice.
It may also cause problems for people if they have two different servers, each with identically
named database, and they want to use one for testing and the other for production. If we ever
find that someone like that exists, we can deal with that case.

I don't think there are any necessary documentation changes here yet since we're recommending
that users answer "no" to the question about setting up a test database.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>