MARC framework test - tests all frameworks
Changed some of the tests so that all frameworks
are tested, not just the first row.

This patch has turned up a reported error in the
unimarc_lecture_pub - items.homebranch
and items.holdinbranch for the WEB framework are
in tab -1 (i.e., ignored), but the check is expecting
them in tab 10.  Not sure if this is a real error or not:
perhaps no items should be created for bibs in that framework;
on the other hand, it is still possible to do so, in which
case the branches need to be set.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>