Wrap up Tags integration - add sysprefs for new installs.
Note that when the user is not logged in, the Tags input box/button will not display.
The user will see a message telling them to log in to add tags.  However, in order to
keep the results display clean, the message will display only once, utilizing the loop
context variable "__first__".  Useful stuff!

I also threw OPACShelfBrowser sysprefs.sql, but it still needs added to updatedatabase.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>