kohabug 2180 - improve display when fines block checkouts
If a patron's fine balance is over the limit set by the
noissuecharges syspref, checkouts are blocked.  However,
this was not made very clear in the interface.

Improved the display by doing the following:

[1] If the fine balance is over the noissuecharges limit, use the "blocker"
styling (i.e., make all of the text red).
[2] Add wording to explicitly signal whether or not the fine balance is
blocking charges.

Documentation changes: new screenshots of blocked checkouts.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>