Fixes bug #2231
In OPAC, when cart is sent by email, the email contains an attachement:
an ISO2709 file with cart biblio records. Most users don't have a clue
what to do with this file.

This patch modifies opac-sendbasket.tmpl to clarify what the attachment is.
It also fix a bug in opac-sendbasket.pl which wasn't reading the template
section containing information about the basket:

  Here is your cart, sent from the Koha Online Catalog... foo

It may be necessary to verify that template text is in 'valid' English...

Signed-off-by: Joshua Ferraro <jmf@liblime.com>