Resolves bug 2250: RFC: New System Preference to Determine which 'Types' and associated icons to display in advanced Search
For documentation, please indicate that as part of profiling,
staff can refer to the AdvancedSearchTypes system preference to
choose where to draw the advanced search 'Types' from. Currently
this is implemented as a choice, between itemtypes and ccodes,
but it's been designed to work with any authorised value so long
as an index exists for searching by that authorised value.

By default, and if this syspref doesn't exist, it will pull from
itemtypes as before.