new automated test for template translatability
This test verifies that the English OPAC and staff templates
can be processed by the string extractor (tmpl_process3.pl)
without error.  If a template contains a parsing error
(at least as far as tmpl_process3.pl is concerned), it may
not be correctly converted when a language translation
is applied.

To run this test, do

prove -v xt/author/translatable-templates.t

Signed-off-by: Joshua Ferraro <jmf@liblime.com>