bug 2254 [2/3]: fix authority types in MARC21 frameworks (DB rev 095)
Corrected two errors found in the authority types used
in the MARC21 frameworks:

* one subfield that used "Meeting Name" instead of "MEETI_NAME"
* various subfields that used "CORP0_NAME" instead of "CORPO_NAME"

The biggest visible consequence of this change is that corporate
name headings (110, 610, 710, 810, etc.) will now link correctly
to their authority records.

In addition, brought the fr-FR MARC21 framework scripts up-to-date
with respect to the changes made to the English versions.

No documentation changes.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>