bug 2171: remove cancel transfer
The "transfers to receive" report included
unused code to delete a transfer.  Since the
"transfers to receive" report is still useful
to give a library a report of items that it should
expect to receive, I am retaining the report.  However,
since the cancel transfer functionality would not
work even if it were uncommented in the template, that
portion is removed.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>