Total overhaul of broken "Items with no checkouts" report.
Previous implementation was a crude copy/paste from a larger report.
CSV output was and is not functional (disabled in interface accordingly).
Output now includes and is sorted by itemcallnumber, because the purpose
of this report is to isolate uncirculating items to cull from the shelves.
Output also includes a summary at the top, with links to individual tables
below when broken out by a grouping.  Also included in the feedback but
default styled to display="none" are the actual SQL queries.  This is
incredibly useful for debugging and should probably be done in all reports.

Yet TODO: allow date range for checkouts and instead of
"no checkouts" (=0) allow the query to filter based on =x, >x, <x, etc.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>