bug 2329: removing unused RSS transport type - DB version 098.
I didn't implement a way for patrons to get their advance notices and other
messages through RSS, so I'm removing it from the list of available transport types.
This will cause the "RSS" column in the patron messasging pages to be
unavailable. When we implement RSS, we can add it back in and the checkboxes will return.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>