Adds a stylesheet for Zebra to do an on-the-fly conversion of the Koha 952 item fields to the MARC21 standard 852 location fields:
Please add the following to the documentation in a section explaining
the setup of a SRU/Z39.50 server. You can also pull info from:

http://wiki.koha.org/doku.php?id=en:development:z3950sru

Modifying Output of Your Records On-the-fly

Koha supports modifying the output of your MARC records for public-facing
Z39.50 and SRU servers. You can set up custom mappings to convert fields such
as 952 (Koha items) to 852 (MARC21 location fields). For example here would
be a popular mapping:
Name:		Koha Field		Mapping in MARC21
---------------------------
Location: 	952 $b => 		852 $b
Barcode: 	952 $p => 		852 $p
Price: 		952 $v => 		852 $r
Item type: 	952 $y => 		852 $w
Notes:		952 $z => 		852 $z

This would make the output from the SRU server suitable for a
resource sharing network that used the above convention for location
information. To add this configuration to your Zebra, add the following
block of code in the koha-conf.xml in the <retrievalinfo> area:

<retrieval syntax="xml" name="marc21location">
         <backend syntax="usmarc" name="F">
           <marc inputformat="marc" outputformat="marcxml"
                 inputcharset="utf-8"/>
           <xslt stylesheet="/path/to/koha/koha-tmpl/intranet-tmpl/prog/en/xslt/MARC21slim9522852.xsl"/>
         </backend>
 </retrieval>

The retrieval set name (in this case, marc21location) is arbitrary and
should be set as required by your resource sharing network, so long as it won't
conflict with internal usage within Koha itself. It's the retrieval set name
that is used to trigger the server to respond and parse the record through the
stylesheet and apply the mappings.

In SRU, this is set as follows:

http://example.com:210/biblios?version=1.1&operation=searchRetrieve&query=it&startRecord=1&maximumRecords=20&recordSchema=marc21location

In Z39.50 it's the elementset name construct.