bug 2335: adding SMSSendDriver system preference - DB update to version 3.00.00.105
Previously, the SMSSendDriver system preference was set as a local use system
preference. This patch makes it an official system preference under the "Patron" tab.
This system preference determines which SMS::Send driver is used to send SMS messages.

I have attempted to take care to not overwrite this system preference if it has already
been set.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>