Bug 2084 [1/2]: adding warning to overdues report
This patch adds a warning to the staff interface on the circulation page
to indicate that the overdues report is resource intensive on large installations.

This is considered a stopgap solution. Ideally, this report should be changed in order
to make it usable even for large installations.

Documentation note: Although no functionality has changed, any screenshots of the
circulation page on the staff interface should be updated.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>